refactor(Prize): Prize 패키지 위치 변경
entity에서 configuration으로 이동했습니다.
